Clock

Description

Handling time without dates.

Documentation

type Hour = Integer

A number of hours.

type Min = Integer

A number of minutes.

newtype Clock

A Clock represents a dateless time.

Constructors

Clock

Fields

minutes ∷ Integer

The number of minutes, floored to the day.

fromHourMin ∷ Hour → Min → Clock

Given a number of hours and a number of minutes, return a Clock representing that time, floored to the day, e.g. toString (fromHourMin 24 42) yields "00:42".

fromString ∷ String → Clock

Given a string of the form, hh:mm, return a Clock such that calling toString on the Clock returns the given string.

toString ∷ Clock → String

Given a Clock, return a string representation of the form, hh:mm.

day ∷ Min

The number of minutes in a day, 1440.

hourToMin ∷ Hour → Min

Given a number of hours, return the equivalent number of minutes.