Grains

Description

Calculating the number of grains of wheat on a chessboard, given that the number on each square doubles.

Documentation

square :: Integer -> Integer

Given the number of a square on a chess board, where one grain is placed on the first square and the number of grains on each subsequent square doubles, returns the number of grains of wheat on a given square.

total :: Integer

Returns the total number of grains on the entire chessboard.

Recursive Alternatives

square' :: Integer -> Integer
square' 1 = 1
square' x = 2 * square' (x - 1)

total' :: Integer
total'    = sum (map square' [1..64])