Sublist

Description

Determining if a list is a sublist of another list.

Documentation

data Sublist

Result of comparing two lists to determine if the first is a sublist of the second.

Constructors

Equal
The two lists are equal.

Sublist
The first list is a fully contained, in order, in the second list.

Superlist
The first list fully contains, in order, the second list.

Unequal
Neither list fully contains the other.

sublist :: Eq a => [a] -> [a] -> Sublist

Given two lists, returns a Sublist based on the first list compared to the second.