SumOfMultiples

Description

Calculating the sum of multiples.

Documentation

sumOfMultiplesDefault :: Integral a => a -> a

Given a number, returns the sum of all the multiples of 3 or 5 (sumOfMultiples [3,5]) up to but not including that number.

sumOfMultiples :: Integral a => [a] -> a -> a

Given a list of numbers and a number, returns the sum of all the multiples of any of the numbers in the list up to but not including that number.

fromOneToExclusive :: Integral a => a -> [a]

Given a number, x, returns the list of numbers from 1 to x - 1.

anyDivides :: Integral a => [a] -> [a] -> [a]

Given two lists of numbers, xs and ys, returns a possibly empty list of every y in ys such that some x in xs divides y.

List comprehension:

anyDivides xs ys = [y | y <- ys, any (`divides` y) xs]

LaTeX:

{ y : y in Y quad land quad exists , x ; ( x in X land x | y ) ;}

divides :: Integral a => a -> a -> Bool

Given two numbers, returns True if the first divides the second, otherwise False.